refactor maybe_split_groups to User class
